Extract out HttpSecurityConfiguration
Fixes gh-4394
